Move Rpc from Producer namespace
