+++ Source/EasyNetQ.IntegrationTests/Utils/BusExtensions.cs
@@ -2,8 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
-using EasyNetQ.FluentConfiguration;
-using EasyNetQ.Producer;
 using EasyNetQ.Scheduling;
 
 namespace EasyNetQ.IntegrationTests.Utils
