+++ Source/EasyNetQ.Tests/AutoSubscriberTests/When_autosubscribing_async_with_subscription_configuration_action.cs
@@ -1,8 +1,6 @@
 ﻿// ReSharper disable InconsistentNaming
 using EasyNetQ.AutoSubscribe;
-using EasyNetQ.FluentConfiguration;
 using EasyNetQ.Internals;
-using EasyNetQ.Producer;
 using FluentAssertions;
 using NSubstitute;
 using System;
