+++ Source/EasyNetQ.Tests/AutoSubscriberTests/When_autosubscribing_with_subscription_configuration_attribute_no_expires.cs
@@ -1,7 +1,5 @@
 using EasyNetQ.AutoSubscribe;
-using EasyNetQ.FluentConfiguration;
 using EasyNetQ.Internals;
-using EasyNetQ.Producer;
 using FluentAssertions;
 using NSubstitute;
 using System;
