+++ Source/EasyNetQ.Tests/FluentConfiguration/SubscriptionConfigurationTests.cs
@@ -1,5 +1,4 @@
-﻿using EasyNetQ.FluentConfiguration;
-using Xunit;
+﻿using Xunit;
 
 namespace EasyNetQ.Tests.FluentConfiguration
 {
