+++ Source/EasyNetQ/IBus.cs
@@ -1,5 +1,4 @@
 using System;
-using EasyNetQ.Producer;
 using EasyNetQ.Scheduling;
 
 namespace EasyNetQ
