+++ Source/EasyNetQ/IRpc.cs
@@ -1,15 +1,13 @@
 ﻿using System;
 using System.Threading;
 using System.Threading.Tasks;
-using EasyNetQ.FluentConfiguration;
 using EasyNetQ.Internals;
 
-namespace EasyNetQ.Producer
+namespace EasyNetQ
 {
     /// <summary>
     ///     An RPC style request-response pattern
     /// </summary>
-    //TODO Need support of non-generic overloads
     public interface IRpc : IDisposable
     {
         /// <summary>
