+++ Source/EasyNetQ/RabbitBus.cs
@@ -1,4 +1,3 @@
-using EasyNetQ.Producer;
 using EasyNetQ.Scheduling;
 
 namespace EasyNetQ
